refactor(voice): move ConversationRelay logic from VoiceChannel onto ConversationRelayProvider
VoiceChannel owned every piece of ConversationRelay-specific behavior
directly: the WebSocket setup/prompt/interrupt protocol loop, outbound
call initiation, and the ConversationRelay callback webhook (TwiML
building was already extracted to twiml.py in an earlier PR). Move all
of it onto ConversationRelayProvider — VoiceChannel now only builds the
provider and delegates through thin wrapper methods.

Moved as-is — method bodies are unchanged apart from the necessary
self.X -> self.channel.X renames for functionality VoiceChannel still
owns (TAC, conversation bookkeeping, Calls API client).
_merge_call_options/_build_call_kwargs move in full since they're only
used by outbound, and outbound support itself is provider-specific — no
reason to keep them on VoiceChannel for a hypothetical second provider.

Also renames handle_conversation_relay_callback -> handle_twilio_provider_callback:
this webhook is Twilio's ConversationRelay-specific mechanism, but Media
Streams has its own independent statusCallback equivalent, so the
channel-facing name shouldn't bake in "ConversationRelay" — VoiceProvider
declares the neutral no-op default, ConversationRelayProvider implements it.

Left on VoiceChannel unchanged: Calls-API webhook handling (on_call_status/
on_amd/on_recording and their event handlers), _get_twilio_client, end_call,
get_conversation_session_by_call_sid, process_webhook — all genuinely
provider-agnostic.

No public API changes. Tests updated only at the mechanical level: private
provider methods are now reached via channel._provider, and the channel
argument no longer needs threading through since VoiceProvider now stores
it at construction.
